+++ config.php
@@ -1,0 +1,8 @@
+<?php
+
+define('MYSQL_SERVER', 'localhost');
+define('MYSQL_USER', 'root');
+define('MYSQL_PASSWORD', '');
+define('MYSQL_DB', 'WD05-filmoteka-Grigorovich');
+
+?>
