+++ проверка ошибок POST.php.bak
@@ -1,0 +1,10 @@
+	// Обработка ошибок
+	if ( trim($_POST['title']) == '' ) {
+		$errors[] = '<div class="notify notify--error mb-20">Название фильма не может быть пустым.</div>';
+	}
+	if ( trim($_POST['genre']) == '' ) {
+		$errors[] = '<div class="notify notify--error mb-20">Жанр фильма не может быть пустым.</div>';
+	}
+	if ( trim($_POST['year']) == '' ) {
+		$errors[] = '<div class="notify notify--error mb-20">Год выхода фильма не может быть пустым.</div>';
+	}
